# Node does not need to subclass ABC