@tinyb0t 2023-09-06
Changes made in `tinygrad/`:
```
------------------------------------------------------------
files                             insertions       deletions
------------------------------------------------------------
tinygrad/shape/symbolic.py                 5               5
------------------------------------------------------------
lines added in the tinygrad folder: 0
```